Primeira reorganização da estrutura do projeto em pacotes e classes
