+++ .gitignore
@@ -1,2 +1,1 @@
 *.bak
-*.class
