+++ Problema01.java
@@ -13,43 +13,30 @@
 // import javax.xml.transform.dom.DOMSource;
 // import javax.xml.transform.stream.StreamResult;
 
+import logica.Calculadora;
+import logica.Impressora;
+import logica.LeitorGravador;
+
 class Problema01 {
 
-  public static void backup( String arquivoOriginal, String arquivoBackup ) {
-  }
-
-  public static MeuContainerEscolhido lerAlunosDoArquivoXml( String filename ) throws Exception {
-  }
-
-  public static float calcularMediaGeral( MeuContainerEscolhido alunos ) {
-  }
-
-  public static float calcularNotaMinima( MeuContainerEscolhido alunos ) {
-  }
-
-  public static float calcularNotaMaxima( MeuContainerEscolhido alunos ) {
-  }
-
-  public static int calcularQuantidadeDeAlunos( MeuContainerEscolhido alunos ) {
-  }
-
-  public static void atualizarResumo( String filename, float mediaGeral, float notaMinima, float notaMaxima, int quantidadeDeAlunos ) throws Exception {
-  }
-
-  public static void imprimirTabelaDeAlunos( MeuContainerEscolhido alunos ) {
-  }
+  /*Todas as funções inicialmente colocadas aqui foram realocadas para classes e pacotes
+    mais descritivos e mais organizados. As chamadas são feitas por objetos na classe main().
+   */
 
   public static void main( String[] args ) {
+    Calculadora calculadora = new Calculadora();
+    LeitorGravador leitorGravador = new LeitorGravador();
+    Impressora impressora = new Impressora();
     try {
       String filename = "alunos.xml";
       backup( filename, filename + ".bak" );
-      MeuContainerEscolhido alunos = lerAlunosDoArquivoXml( filename );
-      float mediaGeral = calcularMediaGeral( alunos );
-      float notaMaxima = calcularNotaMaxima( alunos );
-      float notaMinima = calcularNotaMinima( alunos );
-      int quantidadeDeAlunos = calcularQuantidadeDeAlunos( alunos );
-      atualizarResumo( filename, mediaGeral, notaMinima, notaMaxima, quantidadeDeAlunos );
-      imprimirTabelaDeAlunos( alunos );
+      MeuContainerEscolhido alunos = leitorGravador.lerAlunosDoArquivoXml( filename );
+      float mediaGeral = calculadora.calcularMediaGeral( alunos );
+      float notaMaxima = calculadora.calcularNotaMaxima( alunos );
+      float notaMinima = calculadora.calcularNotaMinima( alunos );
+      int quantidadeDeAlunos = calculadora.calcularQuantidadeDeAlunos( alunos );
+      leitorGravador.atualizarResumo( filename, mediaGeral, notaMinima, notaMaxima, quantidadeDeAlunos );
+      impressora.imprimirTabelaDeAlunos( alunos );
     } catch( Exception e ) {
       e.printStackTrace();
       System.out.println( e );
