+++ logica/Calculadora.java
@@ -1,0 +1,16 @@
+package logica;
+
+public class Calculadora {
+
+    public static float calcularMediaGeral( MeuContainerEscolhido alunos ) {
+    }
+
+    public static float calcularNotaMinima( MeuContainerEscolhido alunos ) {
+    }
+
+    public static float calcularNotaMaxima( MeuContainerEscolhido alunos ) {
+    }
+
+    public static int calcularQuantidadeDeAlunos( MeuContainerEscolhido alunos ) {
+    }
+}
