+++ logica/Impressora.java
@@ -1,0 +1,8 @@
+package logica;
+
+public class Impressora {
+
+    public static void imprimirTabelaDeAlunos( MeuContainerEscolhido alunos ) {
+    }
+
+}
