+++ logica/LeitorGravador.java
@@ -1,0 +1,13 @@
+package logica;
+
+public class LeitorGravador {
+
+    public static void backup( String arquivoOriginal, String arquivoBackup ) {
+    }
+
+    public static MeuContainerEscolhido lerAlunosDoArquivoXml( String filename ) throws Exception {
+    }
+
+    public static void atualizarResumo( String filename, float mediaGeral, float notaMinima, float notaMaxima, int quantidadeDeAlunos ) throws Exception {
+    }
+}
